% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{chindex}
\alias{chindex}
\title{Computes the CH index for selecting clustering thresholds}
\usage{
chindex(clusterdf)
}
\arguments{
\item{clusterdf}{Output from `computeclusters`}
}
\description{
Computes the CH index for selecting clustering thresholds
}
