% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{chindices}
\alias{chindices}
\title{Compute CH index over a range of clustering thresholds}
\usage{
chindices(f, clths = seq(0.1, 1.5, length = 20), rescale = TRUE)
}
\arguments{
\item{f}{A fit from `codls`}

\item{clths}{Numeric vector of clustering thresholds}

\item{rescale}{Passed to `computeclusters`}
}
\description{
Compute CH index over a range of clustering thresholds
}
