% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod-package.R
\docType{package}
\name{cod-package}
\alias{cod}
\alias{cod-package}
\title{cod: CodSpeed Benchmark Harness for R}
\description{
Instruments R expressions with CodSpeed's instrument-hooks C library so that benchmark results are recorded when an R process is run under the CodSpeed runner. Provides a thin binding layer over the instrument-hooks C API and a high-level 'test()' interface for writing benchmarks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/cod}
  \item Report bugs at \url{https://github.com/r-lib/cod/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item CodSpeed Technology SAS (Author of the vendored instrument-hooks library) [copyright holder]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
