% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{codbinomial}
\alias{codbinomial}
\title{Fit a genealogical placement GMRF model using maximum likelihood}
\usage{
codbinomial(tr1, logtau = NULL, k = Inf, profcontrol = list(), ...)
}
\arguments{
\item{tr1}{Phylogenetic tree in ape::phylo format}

\item{logtau}{Precision parameter. If NULL, will invoke `tauprofile` to find best value.}

\item{k}{Fits a reduced-rank MRF model if k is an integer < number of nodes in the input tree. This can speed calculation but reduces precision.}

\item{profcontrol}{Optional list of arguments passed to `tauprofile`}

\item{...}{Additional arguments are passed to `mgcv::gam`}
}
\value{
A COD GMRF model fit. Includes the GAM model fit.
}
\description{
Fits the COD GMRF model using the `mgcv::gam` method. Additional arguments can be passed to `gam`; see documentation for that method. Using method="REML" can speed execution by using a constrained maximum likelihood approach. Additionally, an approximate reduced-rank MRF model can be fitted by supplying the `k` parameter. 
If tau is not provided, `codls` is also used to optimise this parameter. 
This method is slower than `codls` and is not recommended for trees with more than several hundred samples.
}
\details{
The ML COD GMRF method does not currently support inverse probability weighting of samples. Use `codls` if sample weighting is needed.
}
