% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{codls}
\alias{codls}
\title{Fit a COD GMRF model using weighted least squares}
\usage{
codls(tr1, logtau = NULL, profcontrol = list(), weights = NULL)
}
\arguments{
\item{tr1}{Phylogenetic tree in ape::phylo format}

\item{logtau}{Precision parameter. If NULL, will invoke `tauprofile` to find best value.}

\item{profcontrol}{Optional list of arguments passed to `tauprofile`}

\item{weights}{An optional vector (named or unnamed) of sample weights for each tip in the input tree}
}
\value{
A GPGMRF model fit
}
\description{
Fit a COD GMRF model using weighted least squares
}
