% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{computeclusters}
\alias{computeclusters}
\title{Compute phylogenetic clusters by cutting tree at branches with large changes in coalescent odds}
\usage{
computeclusters(f, clth, rescale = TRUE, includeinternals = TRUE)
}
\arguments{
\item{f}{A model fit from `codls`}

\item{clth}{Numeric threshold change in coalescent log odds}

\item{rescale}{if TRUE (default), coalescent log odds are rescaled (mean zero, unit variance) prior to applying thresholds}

\item{includeinternals}{if TRUE (default), internal nodes are also included in cluster output}
}
\value{
A data frame with cluster asignment for each tip
}
\description{
Compute phylogenetic clusters by cutting tree at branches with large changes in coalescent odds
}
