% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{lbi}
\alias{lbi}
\title{Compute the local branching index (LBI)}
\usage{
lbi(tr, logtau)
}
\arguments{
\item{tr}{An ape::phylo}

\item{logtau}{Smoothing parameter}
}
\value{
Vector of local branching index values for every node in the input tree
}
\description{
Compute the local branching index (LBI)
}
