% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{neffextant}
\alias{neffextant}
\title{Effective number of extant lineages}
\usage{
neffextant(f)
}
\arguments{
\item{f}{Fit from `codls`}
}
\value{
Numeric effective number extant lineages
}
\description{
Effective number of extant lineages
}
