% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{plotclusters}
\alias{plotclusters}
\title{Plots a fit from `codls` and cluster assigment from `computeclusters`}
\usage{
plotclusters(f, clusterdf, ...)
}
\arguments{
\item{A}{`codls` fit}

\item{Output}{data frame from `computeclusters`}
}
\description{
Plots a fit from `codls` and cluster assigment from `computeclusters`
}
