% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{rmsclo}
\alias{rmsclo}
\title{Root mean square coalescent log odds}
\usage{
rmsclo(f)
}
\arguments{
\item{f}{Fit from `codls`}
}
\value{
Numeric RMSCLO
}
\description{
This is a summary statistic that describes the amount of variation in coalescent rates across lineages in a phylogenetic tree  
It is defined as \deqn{ \sqrt{ \sum_i l_i \psi_i^2 / L }  }
where the sum is over all branches i in the tree and weighted by branch length \eqn{l_i} and where  \eqn{ L = \sum_i l_i  }
}
