% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{tauprofile}
\alias{tauprofile}
\title{Evaluate the loss function in the GPEGMRF model across a range of tau (precision parameter) values}
\usage{
tauprofile(
  tr,
  logtaulb = -4,
  logtauub = 35,
  res = 11,
  startpc = 75,
  endpc = 100,
  nobj = 100,
  ipw = NULL
)
}
\arguments{
\item{tr}{A phylogenetic tree in ape::phylo format}

\item{logtaulb}{Lower bound of precision parameteters}

\item{logtauub}{Upper bound of precision parameteters}

\item{res}{Number of tau values to evaluate}

\item{startpc}{The initial per cent of nodes in the tree counting from root to tips where the loss function will be evaluated}

\item{endpc}{The final per cent of nodes in the tree counting from root to tips where the loss function will be evaluated}

\item{nobj}{The integer number of points along the tree where the loss function will be evaluated. If Inf, will use all points between startpc and endpc, but may be slow.}

\item{ipw}{Optional inverse probability weights for each sample}
}
\value{
A data frame containing the loss function evaluated over a range of tau values
}
\description{
Evaluate the loss function in the GPEGMRF model across a range of tau (precision parameter) values
}
