% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{codls}
\alias{codls}
\title{Fit a COD GMRF model using weighted least squares}
\usage{
codls(
  tr1,
  logtau = c(0, NULL),
  profcontrol = list(),
  inverseprobabilityweights = NULL
)
}
\arguments{
\item{tr1}{Phylogenetic tree in ape::phylo format}

\item{profcontrol}{Optional list of arguments passed to `tauprofile`}

\item{inverseprobabilityweights}{An optional vector (named or unnamed) of sample weights for each tip in the input tree}

\item{tau}{Precision parameter. If NULL, will invoke `tauprofile` to find best value.}
}
\value{
A GPGMRF model fit
}
\description{
Fit a COD GMRF model using weighted least squares
}
