% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{codls}
\alias{codls}
\title{Fit a COD GMRF model using weighted least squares}
\usage{
codls(tr1, logtau = NULL, profcontrol = list(), weights = NULL)
}
\arguments{
\item{tr1}{Phylogenetic tree in ape::phylo format}

\item{logtau}{Precision parameter. If NULL, will invoke `tauprofile` to find best value.}

\item{profcontrol}{Optional list of arguments passed to `tauprofile`}

\item{weights}{An optional vector (named or unnamed) of sample weights for each tip in the input tree}
}
\value{
A GPGMRF model fit
}
\description{
Fit a COD GMRF model using weighted least squares
}
\examples{
# A simple example that does not have population structure 
set.seed( 1111 )
tr <- ape::rtree( 100 )
f <- codls(tr)
coef(f) |> head() 
summary(f)
\dontrun{
plot(f)
}

# This example has population structure 
tr0 = rcoal(20); tr0$edge.length <- .01*tr0$edge.length 
tr1 = rcoal(80); 
dx <- (max(node.depth.edgelength( tr1 ))-max(node.depth.edgelength( tr0 )))
tr0$root.edge <- dx
tr <- bind.tree(tr0,tr1, position = dx)
f <- codls(tr)
summary(f) 
\dontrun{
plot(f)
}

}
