% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppe2.R
\name{codml}
\alias{codml}
\title{Fit a genealogical placement GMRF model using maximum likelihood}
\usage{
codml(tr1, logtau = c(0, NULL), profcontrol = list(), ...)
}
\arguments{
\item{tr1}{Phylogenetic tree in ape::phylo format}

\item{profcontrol}{Optional list of arguments passed to `tauprofile`}

\item{...}{Additional arguments are passed to `optim`}

\item{tau}{Precision parameter. If NULL, will invoke `tauprofile` to find best value.}
}
\value{
A GPGMRF model fit
}
\description{
This method optimises the sequential-bernoulli likelihood of a COD GMRF model using gradient descent and using `codls` to find initial conditions. 
If tau is not provided, `codls` is also used to optimise this parameter. 
This method is slower than `codls` and is not recommended for trees with more than several hundred samples.
}
\details{
The ML COD GMRF model does not support inverse probability weighting of samples. Use `codls` if sample weighting is needed.
}
